eq_proj/requirements.txt
asgiref==3.4.1
Django==3.2.6
gunicorn==20.1.0
Pillow==8.3.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0
pytz==2021.1
sqlparse==0.4.1
